+++ lib/oneview-sdk/resource/api600/synergy/interconnect.rb
@@ -10,28 +10,13 @@
 # language governing permissions and limitations under the License.
 
 require_relative '../c7000/interconnect'
-require_relative '../../api300/synergy/interconnect'
+require_relative '../../api500/synergy/interconnect'
 
 module OneviewSDK
   module API600
     module Synergy
       # Interconnect resource implementation on API600 Synergy
       class Interconnect < OneviewSDK::API600::C7000::Interconnect
-
-        # Retrieves the interconnect link topologies
-        # @param [OneviewSDK::Client] client The client object for the OneView appliance
-        # @return [Array] All the Interconnect Link Topologies
-        def self.get_link_topologies(client)
-          OneviewSDK::API300::Synergy::Interconnect.get_link_topologies(client)
-        end
-
-        # Retrieves the interconnect link topology with the name
-        # @param [OneviewSDK::Client] client The client object for the OneView appliance
-        # @param [String] name Switch type name
-        # @return [Array] Switch type
-        def self.get_link_topology(client, name)
-          OneviewSDK::API300::Synergy::Interconnect.get_link_topology(client, name)
-        end
       end
     end
   end