+++ spec/integration/resource/api600/c7000/interconnect/create_spec.rb
@@ -11,11 +11,11 @@
 
 require 'spec_helper'
 
-klass = OneviewSDK::API500::C7000::Interconnect
+klass = OneviewSDK::API600::C7000::Interconnect
 RSpec.describe klass, integration: true, type: CREATE do
-  let(:current_client) { $client_500 }
+  let(:current_client) { $client_600 }
   let(:interconnect_name) { INTERCONNECT_2_NAME }
   let(:interconnect_type) { 'HP VC FlexFabric-20/40 F8 Module' }
 
-  include_examples 'InterconnectCreateExample', 'integration api500 context', 500
+  include_examples 'InterconnectCreateExample', 'integration api600 context', 600
 end